requirements.txt
# PyTorch CUDA wheels (needed for +cu118 builds)
--extra-index-url https://download.pytorch.org/whl/cu118

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio Processing
ffmpeg-python==0.2.0
librosa==0.10.1
soundfile==0.12.1
pyannote.audio==3.1.1
torchaudio==2.1.0+cu118
scipy==1.11.4

# ML Framework
torch==2.1.0+cu118
transformers==4.35.2
accelerate==0.24.1
datasets==2.14.7

# Speech Models
speechbrain==0.5.16
openai-whisper==20231117
whisperx==3.1.1
# Optional: WeSpeaker is not published on PyPI.
# Install from source if needed (Linux recommended) and comment out below:
# pip install git+https://github.com/wenet-e2e/wespeaker.git
silero==0.4.1

# Indian Language Support
# IndicTrans2 is loaded via Hugging Face transformers (no separate PyPI package needed)
# If you specifically need the legacy 'indictrans' package, install from source instead of PyPI.
# ai4bharat-transliteration is optional and not used directly; omit to avoid heavy deps
# Optional: Indic NLP utilities (not used directly in code)
# indic-nlp-library==0.9.0

# Storage & Serving
minio==7.2.0
tritonclient[all]==2.40.0
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.1
Pillow==10.0.1
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
prometheus-client==0.19.0
